Changed graph representation from dict[int, dict[int, int]] to dict[int, dict[int, float]]
Changed the graph representation to store edge weights as floats. Additionally, altered the graph generators to produce floating point edge weights.

+++ src/fineman/betweenness_reduction.py
@@ -2,7 +2,7 @@ from math import ceil, log
 import random as rand
 from src.fineman.helper_functions import b_hop_sssp, b_hop_stsp, super_source_bfd
 
-def betweenness_reduction(graph: dict[int, dict[int, int]], neg_edges, tau, beta, c, seed = None):
+def betweenness_reduction(graph: dict[int, dict[int, float]], neg_edges, tau, beta, c, seed = None):
     if (beta < 1) or (tau < 1) or (tau > len(graph)) or (c <= 1):
         raise ValueError("Invalid parameter")
 
@@ -27,7 +27,7 @@ def betweenness_reduction(graph: dict[int, dict[int, int]], neg_edges, tau, beta
     return super_source_bfd(h_graph, h_neg_edges, l, cycleDetection=True)
 
 
-def _construct_h(graph: dict[int, dict[int, int]], T, distances):
+def _construct_h(graph: dict[int, dict[int, float]], T, distances):
     h_graph = {}
     h_neg_edges = set()
